requirements.txt
# Flask web application
Flask==2.3.3
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.21
PyMySQL==1.1.0
mysqlclient==2.2.0

# MQTT Client for IoT devices
paho-mqtt==2.1.0

# Environment variable management
python-dotenv==1.0.0

# Utilities
Werkzeug==2.3.7